+++ packages/frontend/src/components/MkServerSetupWizard.vue
@@ -191,7 +191,6 @@ SPDX-License-Identifier: AGPL-3.0-only
 <script setup lang="ts">
 import { computed, ref } from 'vue';
 import * as Misskey from 'misskey-js';
-import { ROLE_POLICIES } from '@@/js/const.js';
 import MkButton from '@/components/MkButton.vue';
 import MkInput from '@/components/MkInput.vue';
 import * as os from '@/os.js';
