+++ packages/frontend/src/pages/admin/roles.vue
@@ -331,7 +331,7 @@ SPDX-License-Identifier: AGPL-3.0-only
 
 <script lang="ts" setup>
 import { computed, reactive, ref } from 'vue';
-import { ROLE_POLICIES } from '@@/js/const.js';
+import * as Misskey from 'misskey-js';
 import MkInput from '@/components/MkInput.vue';
 import MkFolder from '@/components/MkFolder.vue';
 import MkSwitch from '@/components/MkSwitch.vue';
@@ -353,8 +353,8 @@ const baseRoleQ = ref('');
 
 const roles = await misskeyApi('admin/roles/list');
 
-const policies = reactive<Record<typeof ROLE_POLICIES[number], any>>({});
-for (const ROLE_POLICY of ROLE_POLICIES) {
+const policies = reactive<Record<typeof Misskey.rolePolicies[number], any>>({});
+for (const ROLE_POLICY of Misskey.rolePolicies) {
 	policies[ROLE_POLICY] = instance.policies[ROLE_POLICY];
 }
